% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_luz.R
\name{as_luz_module}
\alias{as_luz_module}
\title{Convert a Scorch Model to a Luz-Compatible Module}
\usage{
as_luz_module(scorch_model)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}
}
\value{
A \code{torch::nn_module} generator suitable for \pkg{luz}.
}
\description{
Convert a Scorch Model to a Luz-Compatible Module
}
\seealso{
Other model training: 
\code{\link{as_torch}()},
\code{\link{fit_scorch}()},
\code{\link{scorch_evaluate}()},
\code{\link{scorch_evaluate_predictions}()},
\code{\link{scorch_predict}()}
}
\concept{model training}
