% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_plotting.R
\name{autoplot.scorch_model}
\alias{autoplot.scorch_model}
\title{Autoplot a Scorch Model}
\usage{
\method{autoplot}{scorch_model}(object, type = c("architecture", "parameters"), ...)
}
\arguments{
\item{object}{A \code{scorch_model} object.}

\item{type}{Plot type: \code{"architecture"} or \code{"parameters"}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Autoplot a Scorch Model
}
