% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{calc_dim}
\alias{calc_dim}
\title{Calculate Dimensions of a Tensor}
\usage{
calc_dim(tensor)
}
\arguments{
\item{tensor}{A tensor object.}
}
\value{
A string representing the dimensions of the tensor, excluding the
batch dimension. If the tensor has only one dimension, returns 1.
}
\description{
Calculates the dimensions of a given tensor, returning a formatted string.
}
\examples{

input = mtcars |> as.matrix() |> torch::torch_tensor()

calc_dim(input)
}
