% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{create_dataloader}
\alias{create_dataloader}
\title{Create a new dataloader for building a model}
\usage{
create_dataloader(
  input,
  output,
  aux = NULL,
  name = "dl",
  batch_size = 32,
  shuffle = TRUE
)
}
\arguments{
\item{input}{A matrix with input data (usually your predictors)}

\item{output}{A matrix with output data (usually what you are predicting)}

\item{aux}{An (optional) matrix with auxiliary training data (usually for training/loss)}

\item{batch_size}{The size of the batch you want to load}

\item{shuffle}{Whether to randomly shuffle the samples}
}
\value{
The dataloader with the specified batch size as a scorch_dataloader
}
\description{
Create a new dataloader for building a model
}
\examples{

input = mtcars |> as.matrix()
output = mtcars |> as.matrix()
dl = create_dataloader(input,output,batch_size=2)
}
