% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{create_scorch_dataloader_class}
\alias{create_scorch_dataloader_class}
\title{Create a Scorch Dataloader Class}
\usage{
create_scorch_dataloader_class(dl)
}
\arguments{
\item{dl}{A dataloader object.}
}
\value{
The input dataloader object with the class attribute set to include
'scorch_dataloader'.
}
\description{
This function adds the class 'scorch_dataloader' to a given dataloader
object.
}
\examples{

input = mtcars  |> as.matrix() |> torch::torch_tensor()

output = mtcars |> as.matrix() |> torch::torch_tensor()

dl = create_dataloader(input, output, batch_size = 2)

scorch_dataloader <- create_scorch_dataloader_class(dl)

class(scorch_dataloader)
}
