% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_scorch.R
\name{create_scorch_model_class}
\alias{create_scorch_model_class}
\title{Create a Scorch Model Class}
\usage{
create_scorch_model_class(obj)
}
\arguments{
\item{obj}{An object to be converted to a scorch model.}
}
\value{
The input object with the class attribute set to 'scorch_model'.
}
\description{
#' This function creates an object of class 'scorch_model'.
}
\examples{

scorch_model <- create_scorch_model_class(list(scorch_architecture = list()))

class(scorch_model)
}
