% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_tfrecord.R
\name{create_scorch_tfrecord_class}
\alias{create_scorch_tfrecord_class}
\title{Create a scorch TFRecord Class}
\usage{
create_scorch_tfrecord_class(obj)
}
\arguments{
\item{obj}{A list containing tfrecord dataset components.}
}
\value{
The input object with class attribute set to include
\code{"scorch_tfrecord"}.
}
\description{
Adds the class \code{"scorch_tfrecord"} to a given list object.
}
\examples{
\dontrun{
tfrecord <- list(
  input = torch::torch_tensor(matrix(1:12, nrow = 3)),
  output = torch::torch_tensor(c(1, 2, 1)),
  n_samples = 3,
  n_features = 4
)

scorch_tfrecord <- create_scorch_tfrecord_class(tfrecord)
class(scorch_tfrecord)
}
}
