% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{head.scorch_dataloader}
\alias{head.scorch_dataloader}
\title{Create the dataloader head function}
\usage{
\method{head}{scorch_dataloader}(dl, ...)
}
\arguments{
\item{dl}{This is the data loader you want to see}
}
\description{
Create the dataloader head function
}
\examples{

input = mtcars |> as.matrix()
output = mtcars |> as.matrix()
dl = create_dataloader(input,output,batch_size=2)
head(dl)
}
