% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_tfrecord.R
\name{head.scorch_tfrecord}
\alias{head.scorch_tfrecord}
\title{Head Method for scorch TFRecord}
\usage{
\method{head}{scorch_tfrecord}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"scorch_tfrecord"}.}

\item{...}{Additional arguments passed to \code{\link[utils]{head}}.}
}
\value{
A list containing the first elements of the input and output tensors.
}
\description{
Returns the first elements of the input and output data from a
\code{scorch_tfrecord} object.
}
\examples{
\dontrun{
dataset <- scorch_tfrecord(filepaths, output = labels)
head(dataset)
}
}
