% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_read_wfdb_file.R
\name{head.scorch_wfdb}
\alias{head.scorch_wfdb}
\title{Head Method for scorch WFDB}
\usage{
\method{head}{scorch_wfdb}(x, n = 6L, ...)
}
\arguments{
\item{x}{An object of class \code{"scorch_wfdb"}.}

\item{n}{Integer. Number of timepoints (columns) to show. Default is 6
(standard R convention).}

\item{...}{Additional arguments (currently unused).}
}
\value{
If \code{x$success} is \code{TRUE}, a matrix with all leads and the
first \code{n} columns. If \code{x$success} is \code{FALSE}, returns
\code{NULL} with a message.
}
\description{
Returns the first \code{n} timepoints (columns) of the signal matrix from
a \code{scorch_wfdb} object. All signals/leads are shown.
}
\examples{
\dontrun{
ecg <- scorch_read_wfdb_file("record.dat", "record.hea")
head(ecg)       # first 6 timepoints
head(ecg, 10)   # first 10 timepoints
}
}
