% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_embeddings.R
\name{IdentityEmbedding}
\alias{IdentityEmbedding}
\title{Identity Positional Embedding}
\usage{
IdentityEmbedding()
}
\value{
An \code{IdentityEmbedding} object.
}
\description{
The \code{IdentityEmbedding} module applies an identity transformation,
effectively passing the input through without modification.
}
\examples{

x <- torch::torch_tensor(1:10)

emb <- IdentityEmbedding()

emb(x)

}
