% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_embeddings.R
\name{LinearEmbedding}
\alias{LinearEmbedding}
\title{Linear Positional Embedding}
\usage{
LinearEmbedding(size, scale = 1)
}
\arguments{
\item{size}{The dimension of the embedding.}

\item{scale}{A scaling factor applied to the input. Default is 1.0.}
}
\value{
A \code{LinearEmbedding} object.
}
\description{
The \code{LinearEmbedding} module implements a simple linear positional embedding,
which scales the input and adds a dimension for the embedding.
}
\examples{

x <- torch::torch_tensor(1:10)

emb <- LinearEmbedding(8)

emb(x)

}
