% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_diffusion.R
\name{NoiseScheduler}
\alias{NoiseScheduler}
\title{Noise Scheduler for Diffusion Process}
\usage{
NoiseScheduler(
  num_timesteps = 1000,
  beta_start = 1e-04,
  beta_end = 0.02,
  beta_schedule = "linear"
)
}
\arguments{
\item{num_timesteps}{The number of timesteps in the diffusion process.
Default is 1000.}

\item{beta_start}{The starting value of beta for the noise schedul.
Default is 0.0001.}

\item{beta_end}{The ending value of beta for the noise schedule.
Default is 0.02.}

\item{beta_schedule}{The schedule type for beta values, either "linear" or
"quadratic". Default is "linear".}
}
\value{
A \code{NoiseScheduler} object with methods for managing the diffusion
process.
}
\description{
The \code{NoiseScheduler} module handles the scheduling of noise addition in the
diffusion process. It supports linear and quadratic beta schedules and
provides methods for reconstructing the original signal, computing posterior
means, and adding noise to samples.
}
\references{
\url{https://github.com/tanelp/tiny-diffusion}
}
