% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{print.scorch_dataloader}
\alias{print.scorch_dataloader}
\title{Create the dataloader print function}
\usage{
\method{print}{scorch_dataloader}(dl)
}
\arguments{
\item{dl}{This is the data loader you want to see}
}
\description{
Create the dataloader print function
}
\examples{

input = mtcars |> as.matrix()
output = mtcars |> as.matrix()
dl = create_dataloader(input,output)
dl
}
