% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_scorch.R
\name{print.scorch_nn_module}
\alias{print.scorch_nn_module}
\title{Print Method for Scorch Neural Network Module}
\usage{
\method{print}{scorch_nn_module}(x, ...)
}
\arguments{
\item{x}{An object of class 'scorch_nn_module'.}

\item{...}{Additional arguments to be passed to the print function.}
}
\description{
This function defines the print method for objects of class
'scorch_nn_module'.
}
\examples{

model <- list(children = list(modules = "module details"))

model$children$modules <- "modules: 123\nlayer1 #64\nlayer2 #32"

class(model) <- c("scorch_nn_module", class(model))

print(model)
}
