% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_tfrecord.R
\name{print.scorch_tfrecord}
\alias{print.scorch_tfrecord}
\title{Print Method for scorch TFRecord}
\usage{
\method{print}{scorch_tfrecord}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"scorch_tfrecord"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Prints a summary of a \code{scorch_tfrecord} object.
}
\examples{
\dontrun{
dataset <- scorch_tfrecord(metadata$filepaths, output = metadata$labels)
print(dataset)
}
}
