% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_read_wfdb_file.R
\name{print.scorch_wfdb}
\alias{print.scorch_wfdb}
\title{Print Method for scorch WFDB}
\usage{
\method{print}{scorch_wfdb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"scorch_wfdb"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Prints a summary of a \code{scorch_wfdb} object.
}
\examples{
\dontrun{
ecg <- scorch_read_wfdb_file("record.dat", "record.hea")
print(ecg)
}
}
