\name{reexports}
\alias{reexports}
\alias{autoplot}
\title{Objects Re-Exported From Other Packages}
\description{
These objects are imported from other packages and re-exported by scorcher.
}
\details{
\describe{
  \item{\code{autoplot}}{The \pkg{ggplot2} S3 plotting generic used by
  \code{\link{autoplot.scorch_model}} and \code{\link{autoplot.scorch_run}}.}
}
}
\keyword{internal}
