% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_diffusion.R
\name{scorch_2d_diffusion_forward}
\alias{scorch_2d_diffusion_forward}
\title{Helper for Forward Pass for Scorch 2D Diffusion Model}
\usage{
scorch_2d_diffusion_forward(model, input, timesteps)
}
\arguments{
\item{model}{The Scorch model being used.}

\item{input}{The input data, expected to be a tensor with two dimensions.}

\item{timesteps}{The current timesteps in the diffusion process.}
}
\value{
The processed input tensor after embedding and concatenation.
}
\description{
The \code{scorch_2d_diffusion_forward} function defines the additional steps for
the forward pass of the Scorch 2D diffusion model. It processes the input
data and time step into embeddings and concatenates them for further
processing in the model.
}
