% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_diffusion.R
\name{scorch_2d_diffusion_init}
\alias{scorch_2d_diffusion_init}
\title{Helper to Initialize Scorch 2D Diffusion Model}
\usage{
scorch_2d_diffusion_init(
  model,
  emb_size = 128,
  time_emb = "sinusoidal",
  input_emb = "sinusoidal",
  scale = 1
)
}
\arguments{
\item{model}{The Scorch model to initialize.}

\item{emb_size}{The size of the embeddings. Default is 128.}

\item{time_emb}{The type of embedding for time steps, options include
"sinusoidal", "linear", "learnable", "zero", and "identity".
Default is "sinusoidal".}

\item{input_emb}{The type of embedding for inputs, options include
"sinusoidal", "linear", "learnable", "zero", and "identity".
Default is "sinusoidal".}

\item{scale}{The scaling factor for input embeddings. Default is 1.0.}
}
\value{
None. The function modifies the \code{model} in place.
}
\description{
The \code{scorch_2d_diffusion_init} function can be passed to \code{compile_scorch}
and provides additional functionality for initializing a Scorch 2D diffusion
model by setting up positional embeddings for the input and time steps.
}
