% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_diffusion.R
\name{scorch_2d_diffusion_train}
\alias{scorch_2d_diffusion_train}
\title{Helper for Batch Preprocessing for Training in Scorch 2D Diffusion Model}
\usage{
scorch_2d_diffusion_train(batch, noise_scheduler, ...)
}
\arguments{
\item{batch}{A batch of data containing inputs.}

\item{noise_scheduler}{The \code{NoiseScheduler} object used for adding noise.}

\item{...}{Additional arguments for customization.}
}
\value{
A list containing the noisy inputs, the noise added, and the
timesteps used.
}
\description{
The \code{scorch_2d_diffusion_train} function prepares a batch of data for
training by adding noise to the inputs according to the current timestep and
generating the corresponding noisy outputs.
}
