% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_run.R
\name{scorch_audit}
\alias{scorch_audit}
\title{Audit a Scorch Run}
\usage{
scorch_audit(run)
}
\arguments{
\item{run}{A \code{scorch_run} object.}
}
\value{
A tibble with audit checks.
}
\description{
Audit a Scorch Run
}
\seealso{
Other reproducibility: 
\code{\link{scorch_compare_runs}()},
\code{\link{scorch_load_run}()},
\code{\link{scorch_save_run}()},
\code{\link{scorch_snapshot}()}
}
\concept{reproducibility}
