% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_run.R
\name{scorch_compare_runs}
\alias{scorch_compare_runs}
\title{Compare Scorch Runs}
\usage{
scorch_compare_runs(...)
}
\arguments{
\item{...}{\code{scorch_run} objects.}
}
\value{
A tibble with one row per run.
}
\description{
Compare Scorch Runs
}
\seealso{
Other reproducibility: 
\code{\link{scorch_audit}()},
\code{\link{scorch_load_run}()},
\code{\link{scorch_save_run}()},
\code{\link{scorch_snapshot}()}
}
\concept{reproducibility}
