% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_embeddings.R
\name{scorch_embedding}
\alias{scorch_embedding}
\title{Add an Embedding Layer to a Scorch Model}
\usage{
scorch_embedding(scorch_model, size = 128, type = "sinusoidal", ...)
}
\arguments{
\item{scorch_model}{A scorch model object to which an embedding layer will
be added.}

\item{size}{An integer specifying the size of the embedding. Default is 128.}

\item{type}{A character string specifying the type of positional embedding
to use. Options include "sinusoidal", "linear", "learnable", "zero", and
"identity". Default is "sinusoidal".}

\item{...}{Additional arguments passed to the embedding function. These
arguments are specific to the embedding type and can include parameters
like \code{scale} for the sinusoidal embedding.}
}
\value{
The modified Scorch model with the embedding layer added to its
architecture.
}
\description{
The \code{scorch_embedding} function adds a positional embedding layer to the
given Scorch model. This embedding can be one of "sinusoidal", "linear",
"learnable", "zero", or "identity".
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg, batch_size = 16)

scorch_model <- dl |>

  initiate_scorch() |>

  scorch_embedding(size = 128, type = "sinusoidal", scale = 1)

}
