% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_plotting.R
\name{scorch_graph_data}
\alias{scorch_graph_data}
\title{Return Scorch Graph Plot Data}
\usage{
scorch_graph_data(scorch_model)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}
}
\value{
A list with \code{nodes} and \code{edges} tibbles.
}
\description{
Return Scorch Graph Plot Data
}
\seealso{
Other model construction: 
\code{\link{as_torch_module}()},
\code{\link{compile_scorch}()},
\code{\link{initiate_scorch}()},
\code{\link{plot_scorch_model}()},
\code{\link{print.scorch_model}()},
\code{\link{scorch_add_skip}()},
\code{\link{scorch_attention}()},
\code{\link{scorch_batchnorm}()},
\code{\link{scorch_block}()},
\code{\link{scorch_concat}()},
\code{\link{scorch_conv_transpose}()},
\code{\link{scorch_dropout}()},
\code{\link{scorch_flatten}()},
\code{\link{scorch_function}()},
\code{\link{scorch_input}()},
\code{\link{scorch_layer}()},
\code{\link{scorch_output}()},
\code{\link{scorch_pool}()},
\code{\link{scorch_spec}()},
\code{\link{scorch_transformer_encoder}()},
\code{\link{validate_scorch_graph}()}
}
\concept{model construction}
