% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_huggingface.R
\name{scorch_hf_base_module}
\alias{scorch_hf_base_module}
\title{Create an nn_module Constructor for HF Feature-Extraction}
\usage{
scorch_hf_base_module(model_id, api_key = Sys.getenv("HF_API_KEY"))
}
\arguments{
\item{model_id}{HF repo ID to load feature-extraction from.}

\item{api_key}{API token for HF (default from Sys.getenv("HF_API_KEY")).}
}
\value{
A constructor function for an nn_module named "hf_base".
}
\description{
Create an nn_module Constructor for HF Feature-Extraction
}
\examples{

\dontrun{
# Retrieve constructor (no inference yet)

hf_ctor <- scorch_hf_base_module("distilbert-base-uncased")

# Instantiate and get dummy output

mod <- hf_ctor()

out <- mod()  # dummy tensor
}

}
