% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_run.R
\name{scorch_load_run}
\alias{scorch_load_run}
\title{Load a Scorch Run}
\usage{
scorch_load_run(path)
}
\arguments{
\item{path}{File path created by \code{scorch_save_run()}.}
}
\value{
A \code{scorch_run} object.
}
\description{
Load a Scorch Run
}
\seealso{
Other reproducibility: 
\code{\link{scorch_audit}()},
\code{\link{scorch_compare_runs}()},
\code{\link{scorch_save_run}()},
\code{\link{scorch_snapshot}()}
}
\concept{reproducibility}
