% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_evaluate.R
\name{scorch_predict}
\alias{scorch_predict}
\title{Predict with a Scorch Model}
\usage{
scorch_predict(scorch_model, input)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{input}{A torch tensor or named list of torch tensors.}
}
\value{
Model predictions.
}
\description{
Predict with a Scorch Model
}
\seealso{
Other model training:
\code{\link{as_torch}()},
\code{\link{fit_scorch}()},
\code{\link{scorch_evaluate}()},
\code{\link{scorch_evaluate_predictions}()}
}
\concept{model training}
