% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_select.R
\name{scorch_select}
\alias{scorch_select}
\title{Select a Dimension from a Scorch Model}
\usage{
scorch_select(scorch_model, dim)
}
\arguments{
\item{scorch_model}{A scorch model object to which the dimension selection
layer will be added.}

\item{dim}{An integer specifying the dimension to select from the input
tensor. This should be a valid dimension index based on the input tensor's
dimensions.}
}
\value{
The modified scorch model with the dimension selection layer
added to its architecture.
}
\description{
The \code{scorch_select} function adds a layer to the scorch model that selects
a specific dimension from the input tensor. This is useful for applying
operations or layers only to a particular dimension of the input tensor.
}
\examples{

# Example usage:
scorch_model <- dl |>
  initiate_scorch() |>
  scorch_select(dim = 1)

}
