% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_embeddings.R
\name{SinusoidalEmbedding}
\alias{SinusoidalEmbedding}
\title{Sinusoidal Positional Embedding}
\usage{
SinusoidalEmbedding(size, scale = 1)
}
\arguments{
\item{size}{The dimension of the embedding.}

\item{scale}{A scaling factor applied to the input before creating the
embedding. Default is 1.0.}
}
\value{
A \code{SinusoidalEmbedding} object.
}
\description{
The \code{SinusoidalEmbedding} module implements sinusoidal positional embeddings.
It scales the input and applies sine and cosine functions to create
embeddings.
}
\examples{

x <- torch::torch_tensor(1:10)

emb <- SinusoidalEmbedding(8)

emb(x)

}
