% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_scorch.R
\name{compile_scorch}
\alias{compile_scorch}
\title{Compile a Scorch Model}
\usage{
compile_scorch(sm)
}
\arguments{
\item{sm}{A scorch model architecture}
}
\value{
A scorch_model object with components (1) nn_module compiled and
ready to train and (2) the scorch_dataloader object
}
\description{
Compile a Scorch Model
}
\examples{

input  <- mtcars |> as.matrix() |> torch::torch_tensor()

output <- mtcars |> as.matrix() |> torch::torch_tensor()

dl <- scorch_create_dataloader(input, output, batch_size = 2)

scorch_model <- dl |> initiate_scorch() |>

  scorch_layer(torch::nn_linear(11,5)) |>

  scorch_layer(torch::nn_linear(5,2)) |>

  compile_scorch()

}
