% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_scorch.R
\name{compile_scorch}
\alias{compile_scorch}
\title{Compile a Scorch Model}
\usage{
compile_scorch(
  sm,
  loss_fn = nn_mse_loss(),
  optimizer_fn = optim_adam,
  optimizer_params = list(lr = 0.001)
)
}
\arguments{
\item{sm}{A \code{scorch_model} object built via \code{initiate_scorch()} and layers.}

\item{loss_fn}{A loss function, e.g., \code{nn_mse_loss()} or named list of
losses for multi-head models.}

\item{optimizer_fn}{Optimizer constructor, e.g., \code{optim_adam}.}

\item{optimizer_params}{Named list of optimizer parameters
(e.g., \code{list(lr = 1e-3)}).}
}
\value{
The same \code{scorch_model} with \code{nn_model}, \code{optimizer}, and \code{loss_fn}
set, and \code{compiled = TRUE}.
}
\description{
Traverse the topology defined in the scorch model graph to build a single
\code{torch::nn_module}. Also attaches the specified loss function and optimizer.
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg, batch_size=16)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt") |>

  scorch_layer("h1", "linear", in_features = 1, out_features = 8) |>

  scorch_output("h1")

sm <- compile_scorch(sm,

  loss_fn = torch::nn_mse_loss(),

  optimizer_fn = torch::optim_adam,

  optimizer_params = list(lr=0.01))

print(sm)

}
