% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_scorch.R
\name{compile_scorch}
\alias{compile_scorch}
\title{Compile a Scorch Model}
\usage{
compile_scorch(sm, init_fn = NULL, forward_fn = NULL, ...)
}
\arguments{
\item{sm}{A scorch model architecture}

\item{init_fn}{An optional function for initializing the model's parameters.
This function takes the model and additional arguments passed via \code{...}.}

\item{forward_fn}{An optional function for customizing the forward pass of
the model. This function takes the model, input data, and additional
arguments passed via \code{...}.}

\item{...}{Additional arguments passed to the \code{init_fn} and \code{forward_fn}.}
}
\value{
A list containing:
\describe{
\item{nn_model}{The compiled \code{scorch} model object, ready for training.}
\item{dl}{The associated scorch dataloader object.}
}
}
\description{
Compile a Scorch Model
}
\examples{

input  <- mtcars |> as.matrix() |> torch::torch_tensor()

output <- mtcars |> as.matrix() |> torch::torch_tensor()

dl <- scorch_create_dataloader(input, output, batch_size = 2)

scorch_model <- dl |> initiate_scorch() |>

  scorch_layer("linear", 11, 5) |>

  scorch_layer("linear", 5, 2) |>

  compile_scorch()

}
