% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{create_dataloader}
\alias{create_dataloader}
\title{Create a New Dataloader for Building a Scorch Model}
\usage{
create_dataloader(input, output, name = "dl", batch_size = 32, shuffle = TRUE)
}
\arguments{
\item{input}{A tensor representing the input data (usually your predictors,
first dimensions is index).}

\item{output}{A tensor representing the output data (usually what you are
predicting, first dimension is index).}

\item{name}{A character string representing the name of the dataloader.
Default is "dl".}

\item{batch_size}{An integer specifying the batch size. Default is 32.}

\item{shuffle}{A logical value indicating whether to shuffle the data.
Default is TRUE.}
}
\value{
The dataloader with the specified batch size as a scorch_dataloader.
}
\description{
This function creates a dataloader for building a scorch model from the
given input and output tensors.
}
\examples{

input = mtcars  |> as.matrix() |> torch::torch_tensor()

output = mtcars |> as.matrix() |> torch::torch_tensor()

dl = create_dataloader(input, output, batch_size = 2)
}
