% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_scorch.R
\name{fit_scorch}
\alias{fit_scorch}
\title{Fit a Scorch Model}
\usage{
fit_scorch(
  scorch_model,
  loss = nn_mse_loss,
  loss_params = list(reduction = "mean"),
  optim = optim_adam,
  optim_params = list(lr = 0.001),
  num_epochs = 10,
  verbose = TRUE
)
}
\arguments{
\item{scorch_model}{A scorch model object}

\item{loss}{A loss function from the torch package}

\item{loss_params}{Parameters to pass to the loss function}

\item{optim}{An optimizer from the torch package}

\item{optim_params}{Parameters to pass to the optimizer}

\item{num_epochs}{The number of epochs to train for}

\item{verbose}{Whether to print loss results at each epoch}
}
\value{
A trained nn_module
}
\description{
Fit a Scorch Model
}
\examples{

input  <- mtcars |> as.matrix() |> torch::torch_tensor()

output <- mtcars |> as.matrix() |> torch::torch_tensor()

dl <- scorch_create_dataloader(input,output,batch_size=2)

scorch_model <- dl |> initiate_scorch() |>

  scorch_layer(torch::nn_linear(11,5)) |>

  scorch_layer(torch::nn_linear(5,2)) |>

  scorch_layer(torch::nn_linear(2,5)) |>

  scorch_layer(torch::nn_linear(5,11)) |>

  compile_scorch() |>

  fit_scorch()

first_batch <- head(dl)

test_output <- scorch_model(first_batch$input)
}
