% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R
\name{head.scorch_dataloader}
\alias{head.scorch_dataloader}
\title{Head Method for Scorch Dataloader}
\usage{
\method{head}{scorch_dataloader}(x, ...)
}
\arguments{
\item{x}{An object of class 'scorch_dataloader'.}

\item{...}{Additional arguments to be passed to the head function.}
}
\value{
A list containing the first elements of the input and output data
from the dataloader.
}
\description{
Defines the head method for objects of class 'scorch_dataloader', returning
the first elements of the input and output data.
}
\examples{

input  <- mtcars |> as.matrix() |> torch::torch_tensor()

output <- mtcars |> as.matrix() |> torch::torch_tensor()

dl <- scorch_create_dataloader(input, output, batch_size = 2)

head(dl)
}
