% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_scorch_model}
\alias{plot_scorch_model}
\title{Plot a Scorch Model Architecture}
\usage{
plot_scorch_model(scorch_model, detailed = FALSE)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{detailed}{Logical; if TRUE, include parameter shapes.}
}
\value{
A \code{grViz} graph object.
}
\description{
Plot a Scorch Model Architecture
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt") |>

  scorch_layer("h1", "linear", in_features = 1, out_features = 4) |>

  scorch_output("h1") |>

  compile_scorch()

plot_scorch_model(sm)

}
