% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_embeddings.R
\name{PositionalEmbedding}
\alias{PositionalEmbedding}
\title{Positional Embedding Wrapper}
\usage{
PositionalEmbedding(size = NULL, type = "sinusoidal", ..., emb_size = NULL)
}
\arguments{
\item{size}{The dimension of the embedding.}

\item{emb_size}{Optional alias for \code{size}. This is accepted for diffusion
examples that mirror tiny-diffusion naming.}

\item{type}{The type of embedding to use, such as "sinusoidal", "linear",
"learnable", "zero", or "identity".}

\item{...}{Additional parameters specific to the chosen embedding type.}
}
\value{
A \code{PositionalEmbedding} object.
}
\description{
The \code{PositionalEmbedding} module is a wrapper that selects the appropriate
type of positional embedding based on the specified type.
}
\examples{

x <- torch::torch_tensor(1:10)

emb <- PositionalEmbedding(8, "sinusoidal")

emb(x)

}
\references{
\url{https://github.com/tanelp/tiny-diffusion}
}
\seealso{
Other embeddings: 
\code{\link{IdentityEmbedding}()},
\code{\link{LearnableEmbedding}()},
\code{\link{LinearEmbedding}()},
\code{\link{SinusoidalEmbedding}()},
\code{\link{ZeroEmbedding}()}
}
\concept{embeddings}
