% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_embeddings.R
\name{PositionalEmbedding}
\alias{PositionalEmbedding}
\title{Positional Embedding Wrapper}
\usage{
PositionalEmbedding(size, type, ...)
}
\arguments{
\item{size}{The dimension of the embedding.}

\item{type}{The type of embedding to use, such as "sinusoidal", "linear",
"learnable", "zero", or "identity".}

\item{...}{Additional parameters specific to the chosen embedding type.}
}
\value{
A \code{PositionalEmbedding} object.
}
\description{
The \code{PositionalEmbedding} module is a wrapper that selects the appropriate
type of positional embedding based on the specified type.
}
\examples{

x <- torch::torch_tensor(1:10)

emb <- PositionalEmbedding(8, "sinusoidal")

emb(x)

}
\references{
\url{https://github.com/tanelp/tiny-diffusion}
}
