% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_scorch.R, R/methods.R
\name{print.scorch_model}
\alias{print.scorch_model}
\title{Print Method for Scorch Model}
\usage{
\method{print}{scorch_model}(x, detailed = FALSE, ...)

\method{print}{scorch_model}(x, detailed = FALSE, ...)
}
\arguments{
\item{x}{A \code{scorch_model} object.}

\item{detailed}{Logical; if TRUE, include parameter shapes.}

\item{...}{Not used.}
}
\value{
Invisibly returns the original object.
}
\description{
This function defines the print method for objects of class 'scorch_model'.
}
\examples{

# input  <- mtcars |> as.matrix() |> torch::torch_tensor()

# output <- mtcars |> as.matrix() |> torch::torch_tensor()

# dl <- scorch_create_dataloader(input, output, batch_size = 2)

# scorch_model <- dl |> initiate_scorch() |>

#   scorch_layer("linear", 11, 5)

# print(scorch_model)

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt") |>

  scorch_layer("h1", "linear", in_features = 1, out_features = 4) |>

  scorch_output("h1") |>

  compile_scorch()

print(sm)

}
