% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_scorch.R
\name{print.scorch_model}
\alias{print.scorch_model}
\title{Print Method for Scorch Model}
\usage{
\method{print}{scorch_model}(x, ...)
}
\arguments{
\item{x}{An object of class 'scorch_model'.}

\item{...}{Additional arguments to be passed to the print function.}
}
\description{
This function defines the print method for objects of class 'scorch_model'.
}
\examples{
input  = mtcars |> as.matrix() |> torch::torch_tensor()

output = mtcars |> as.matrix() |> torch::torch_tensor()

dl = create_dataloader(input, output, batch_size = 2)

scorch_model = dl |> initiate_scorch() |>

  scorch_layer(torch::nn_linear(11, 5))

print(scorch_model)
}
