% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_tfrecord.R
\name{print.scorch_tfrecord}
\alias{print.scorch_tfrecord}
\title{Print Method for Scorch TFRecord}
\usage{
\method{print}{scorch_tfrecord}(x, ...)
}
\arguments{
\item{x}{An object of class 'scorch_tfrecord'.}

\item{...}{Additional arguments to be passed to the print function.}
}
\value{
NULL. This function is called for its side effect of printing
information about the tfrecord dataset.
}
\description{
Defines the print method for objects of class 'scorch_tfrecord', providing
a summary of its features.
}
\examples{
\dontrun{
dataset <- scorch_tfrecord(metadata$filepaths, output = metadata$labels)

print(dataset)
}

}
