% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataloader.R, R/scorch_tfrecord.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{head}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{utils}{\code{\link[utils]{head}}, \code{\link[utils]{head}}}
}}

