% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_layer.R
\name{scorch_add_skip}
\alias{scorch_add_skip}
\title{Add a Skip Connection Node}
\usage{
scorch_add_skip(scorch_model, name, inputs)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{name}{Unique name for the skip node.}

\item{inputs}{Character vector length 2.}
}
\value{
The updated \code{scorch_model}.
}
\description{
Add a Skip Connection Node
}
\examples{

dl <- scorch_create_dataloader(

  input  = list(x = mtcars$hp),
  output = mtcars$mpg
)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("x") |>

  scorch_add_skip("skip", c("x","x"))

print(sm)

}
