% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_layer.R
\name{scorch_attention}
\alias{scorch_attention}
\title{Add a Multi-Head Attention Node}
\usage{
scorch_attention(scorch_model, name, inputs, embed_dim, num_heads, ...)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{name}{Unique name for the attention node.}

\item{inputs}{Character vector c("query", "key", "value").}

\item{embed_dim}{Embedding dimension.}

\item{num_heads}{Number of attention heads.}

\item{...}{Additional args.}
}
\value{
The updated \code{scorch_model}.
}
\description{
Add a Multi-Head Attention Node
}
\examples{

# Need

}
