% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_layer.R
\name{scorch_batchnorm}
\alias{scorch_batchnorm}
\title{Add a BatchNorm1d Layer}
\usage{
scorch_batchnorm(scorch_model, name, inputs, num_features, ...)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{name}{Unique name for the batchnorm layer.}

\item{inputs}{Single upstream node name.}

\item{num_features}{Number of features (channels).}

\item{...}{Additional args.}
}
\value{
The updated \code{scorch_model}.
}
\description{
Add a BatchNorm1d Layer
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt") |>

  scorch_batchnorm("bn", "wt", num_features = 1)

print(sm)

}
