% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_layer.R
\name{scorch_concat}
\alias{scorch_concat}
\title{Add a Concatenation Node to a Scorch Model}
\usage{
scorch_concat(scorch_model, name, inputs, dim = 1)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{name}{Unique name for the concat node.}

\item{inputs}{Character vector of node names to concatenate.}

\item{dim}{Integer dimension to concatenate along (default 1).}
}
\value{
The updated \code{scorch_model}.
}
\description{
Add a Concatenation Node to a Scorch Model
}
\examples{

dl <- scorch_create_dataloader(

  input  = list(hp = mtcars$hp, disp = mtcars$disp),
  output = mtcars$mpg
)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("hp") |>

  scorch_input("disp") |>

  scorch_concat("merged", c("hp","disp"), dim = 2)

print(sm)

}
