% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_layer.R
\name{scorch_dropout}
\alias{scorch_dropout}
\title{Add a Dropout Layer}
\usage{
scorch_dropout(scorch_model, name, inputs, p = 0.5, ...)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{name}{Unique name for the dropout layer.}

\item{inputs}{Single upstream node name.}

\item{p}{Dropout probability (default 0.5).}

\item{...}{Additional args.}
}
\value{
The updated \code{scorch_model}.
}
\description{
Add a Dropout Layer
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt") |>

  scorch_dropout("do", "wt", p = 0.2)

print(sm)

}
