% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_evaluate.R
\name{scorch_evaluate}
\alias{scorch_evaluate}
\title{Evaluate a Scorch Model}
\usage{
scorch_evaluate(scorch_model, input, truth, metric = c("accuracy", "mse"))
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{input}{A torch tensor or named list of torch tensors.}

\item{truth}{A torch tensor of observed outcomes.}

\item{metric}{Evaluation metric. Currently \code{"accuracy"} and
\code{"mse"} are supported.}
}
\value{
A one-row tibble with the metric value.
}
\description{
Evaluate a Scorch Model
}
\seealso{
Other model training:
\code{\link{as_torch}()},
\code{\link{fit_scorch}()},
\code{\link{scorch_evaluate_predictions}()},
\code{\link{scorch_predict}()}
}
\concept{model training}
