% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_evaluate.R
\name{scorch_evaluate_predictions}
\alias{scorch_evaluate_predictions}
\title{Evaluate Predictions}
\usage{
scorch_evaluate_predictions(predictions, truth, metric = c("accuracy", "mse"))
}
\arguments{
\item{predictions}{A torch tensor of model predictions.}

\item{truth}{A torch tensor of observed outcomes.}

\item{metric}{Evaluation metric. Currently \code{"accuracy"} and
\code{"mse"} are supported.}
}
\value{
A one-row tibble with the metric value.
}
\description{
Evaluate Predictions
}
\seealso{
Other model training: 
\code{\link{as_luz_module}()},
\code{\link{as_torch}()},
\code{\link{fit_scorch}()},
\code{\link{scorch_evaluate}()},
\code{\link{scorch_predict}()}
}
\concept{model training}
