% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_function.R
\name{scorch_function}
\alias{scorch_function}
\title{Add a Custom Function to a Scorch Model}
\usage{
scorch_function(scorch_model, func, ...)
}
\arguments{
\item{scorch_model}{A scorch model object to which a custom function
is to be added.}

\item{func}{A function to be added to the scorch model.}

\item{...}{Additional arguments to be passed to the custom function.}
}
\value{
The updated scorch model with the custom function added to its
architecture.
}
\description{
This function adds a custom function to the scorch model architecture.
}
\examples{

custom_function <- function(x, factor) { x * factor }

input  <- mtcars |> as.matrix() |> torch::torch_tensor()

output <- mtcars |> as.matrix() |> torch::torch_tensor()

dl <- scorch_create_dataloader(input, output, batch_size = 2)

scorch_model <- dl |> initiate_scorch() |>

  scorch_layer("linear", 11, 5) |>

  scorch_function(custom_function, factor = 2)
}
