% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_input.R
\name{scorch_input}
\alias{scorch_input}
\title{Register an Input Node in a Scorch Model}
\usage{
scorch_input(scorch_model, name)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{name}{Unique name for the input (character).}
}
\value{
The updated \code{scorch_model}.
}
\description{
Register an Input Node in a Scorch Model
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt")

print(sm$inputs)

}
