% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_layer.R
\name{scorch_layer}
\alias{scorch_layer}
\title{Add a Layer to a Scorch Model}
\usage{
scorch_layer(scorch_model, nn_obj)
}
\arguments{
\item{scorch_model}{A scorch model object to which a layer is to be added.}

\item{nn_obj}{A neural network object representing the layer to be added
to the scorch model.}
}
\value{
The updated scorch model with the new layer added to its
architecture.
}
\description{
This function adds a neural network layer to the scorch model architecture.
}
\examples{

input  <- mtcars |> as.matrix() |> torch::torch_tensor()

output <- mtcars |> as.matrix() |> torch::torch_tensor()

dl <- scorch_create_dataloader(input, output, batch_size = 2)

scorch_model <- dl |> initiate_scorch() |>

  scorch_layer("linear", 11, 5)
}
