% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorch_output.R
\name{scorch_output}
\alias{scorch_output}
\title{Mark Output Nodes in a Scorch Model}
\usage{
scorch_output(scorch_model, outputs)
}
\arguments{
\item{scorch_model}{A \code{scorch_model} object.}

\item{outputs}{Character vector of node names to mark as outputs.}
}
\value{
The updated \code{scorch_model}.
}
\description{
Mark Output Nodes in a Scorch Model
}
\examples{

dl <- scorch_create_dataloader(mtcars$wt, mtcars$mpg)

sm <- dl |>

  initiate_scorch() |>

  scorch_input("wt") |>

  scorch_output("wt")

print(sm$outputs)

}
